cleanup and finalize today birthday code
